+++ libmpeg/source/mpeg-ts-dec.c
@@ -298,7 +298,7 @@ int ts_demuxer_destroy(struct ts_demuxer_t* ts)
 int ts_demuxer_getservice(struct ts_demuxer_t* ts, int program, char* provider, int nprovider, char* name, int nname)
 {
     struct pmt_t* pmt;
-    pmt = pat_find(&ts->pat, program);
+    pmt = pat_find(&ts->pat, (uint16_t)program);
     if(NULL == pmt)
         return -1;
