+++ librtp/source/rtcp.c
@@ -194,7 +194,7 @@ int rtcp_input_rtcp(struct rtp_context *ctx, const void* data, int bytes)
 			break;
 
 		// RFC3550 6.3.3 Receiving an RTP or Non-BYE RTCP Packet (p26)
-		ctx->avg_rtcp_size = ctx->avg_rtcp_size*1.0/16 + r * 15.0/16;
+		ctx->avg_rtcp_size = (int)(ctx->avg_rtcp_size*1.0/16 + r * 15.0/16);
 
 		p += r;
 		bytes -= r;
